requirements.txt
streamlit==1.54.0
pandas==2.3.3
plotly==6.5.2
firecrawl-py==4.18.0
python-dotenv==1.2.1
pydantic==2.12.5
beautifulsoup4==4.14.3
requests==2.32.5
supabase==2.28.0
